+++ package.json
@@ -10,15 +10,16 @@
   },
   "dependencies": {
     "@next/font": "^13.3.0",
-    "@types/node": "18.15.11",
-    "@types/react": "18.0.35",
-    "@types/react-dom": "18.0.11",
+    "@types/node": "^20",
+    "@types/react": "^18",
+    "@types/react-dom": "^18",
     "autoprefixer": "10.4.14",
+    "eslint-config-next": "^14.0.4",
     "eventsource-parser": "^1.0.0",
-    "next": "13.3.0",
+    "next": "^14.0.4",
     "postcss": "8.4.21",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "tailwindcss": "3.3.1",
     "tiktoken": "^1.0.11",
     "typescript": "5.0.4"
